+++ src/action/check/paymentMethod/movieTicket.ts
@@ -10,17 +10,38 @@ import * as PayActionFactory from '../../trade/pay';
 export type IAgent = ActionFactory.IParticipantAsWebApplication | ActionFactory.IParticipantAsPerson;
 export type IPurchaseNumberAuthIn = surfrock.service.auth.factory.IPurchaseNumberAuthIn;
 export type IPurchaseNumberAuthResult = surfrock.service.auth.factory.IPurchaseNumberAuthResult;
-export type IMovieTicket = Omit<IMovieTicketPaymentCard, 'project' | 'serviceOutput'> & {
-    // 最適化(2023-03-03~)
-    serviceOutput: Pick<IMovieTicketServiceOutput, 'reservationFor'>;
-};
+/**
+ * 認証対象のmovieTicket
+ */
+export type IMovieTicket =
+    // 最適化(2024-03-15~)
+    // Omit<IMovieTicketPaymentCard, 'project' | 'serviceOutput'>
+    Pick<
+        IMovieTicketPaymentCard,
+        'accessCode' | 'category' | 'identifier'
+    // | 'serviceType' // 廃止(2024-03-15~)
+    // | 'typeOf' // 廃止(2024-03-15~)
+    >
+    & {
+        // 最適化(2023-03-03~)
+        serviceOutput: Pick<IMovieTicketServiceOutput, 'reservationFor'>;
+    };
 // 最適化(2023-03-03~)
-export interface IPaymentService extends Pick<PayActionFactory.IPaymentService, 'id' | 'typeOf' | 'paymentMethod'> {
+export interface IPaymentService extends Pick<
+    PayActionFactory.IPaymentService, 'id' | 'typeOf'
+// | 'paymentMethod' // 最適化(2024-03-15~)
+> {
     /**
      * 販売者
      */
     seller: { typeOf: OrganizationType.Corporation; id: string };
     movieTickets?: IMovieTicket[];
+    paymentMethod: {
+        /**
+         * 決済方法区分
+         */
+        typeOf: string;
+    };
 }
 export type IObject = IPaymentService[];
 export interface IResult {